7.03, 7.04, 7.06

+++ Mod01/HelloWorld.java
@@ -8,7 +8,7 @@ public class HelloWorld
    public static void main(String[] args)
    {
        
-       int x = 10.0/3;
+       double x = 10.0/3;
